--- src/main/java/com/onetier/retro_togather/controller/request/LoginRequestDto.java
@@ -1,23 +1,0 @@
-package com.onetier.retro_togather.controller.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-/**
- * LoginRequestDto
- */
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginRequestDto {
-
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String password;
-
-}
