--- src/main/java/com/onetier/retro_togather/controller/request/TokenDto.java
@@ -1,22 +1,0 @@
-package com.onetier.retro_togather.controller.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * TokenDto
- */
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TokenDto {
-    private String grantType;
-    private String accessToken;
-    private String refreshToken;
-    private Long accessTokenExpiresIn;
-
-
-}
