--- src/main/java/com/onetier/retro_togather/controller/response/ResponseDto.java
@@ -1,21 +1,0 @@
-package com.onetier.retro_togather.controller.response;
-import com.onetier.retro_togather.domain.Error;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ResponseDto<T> {
-  private boolean success;
-  private T data;
-  private Error error;
-
-  public static <T> ResponseDto<T> success(T data) {
-    return new ResponseDto<>(true, data, null);
-  }
-
-  public static <T> ResponseDto<T> fail(String code, String message) {
-    return new ResponseDto<>(false, null, new Error(code, message));
-  }
-
-}
