--- src/main/java/com/onetier/retro_togather/domain/Error.java
@@ -1,16 +1,0 @@
-package com.onetier.retro_togather.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Error {
-    private String field;
-    private String message;
-
-}
