--- src/main/java/com/onetier/retro_togather/domain/UserDetailsImpl.java
@@ -1,58 +1,0 @@
-package com.onetier.retro_togather.domain;
-
-
-import com.onetier.retro_togather.shared.Authority;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDetailsImpl implements UserDetails {
-    private Member member;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(Authority.ROLE_MEMBER.toString());
-        Collection<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(authority);
-        return authorities;
-    }
-
-    @Override
-    public String getPassword(){
-        return member.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getEmail();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
