--- src/main/java/com/onetier/retro_togather/service/UserDetailsServiceImpl.java
@@ -1,27 +1,0 @@
-package com.onetier.retro_togather.service;
-
-import com.onetier.retro_togather.domain.Member;
-import com.onetier.retro_togather.domain.UserDetailsImpl;
-import com.onetier.retro_togather.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@RequiredArgsConstructor
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String memberId) throws UsernameNotFoundException {
-        Optional<Member> member = memberRepository.findByNickname(memberId);
-        return member
-                .map(UserDetailsImpl::new)
-                .orElseThrow(()-> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-    }
-}
