--- src/main/java/com/onetier/retro_togather/shared/Authority.java
@@ -1,7 +1,0 @@
-package com.onetier.retro_togather.shared;
-
-public enum Authority {
-    ROLE_MEMBER,
-    ROLE_GUEST
-
-}
