+++ src/main/java/com/onetier/retro_together/RetroTogetherApplication.java
@@ -2,7 +2,9 @@ package com.onetier.retro_together;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class RetroTogetherApplication {
