+++ src/main/java/com/onetier/retro_together/configuration/JwtSecurityConfiguration.java
@@ -1,0 +1,30 @@
+package com.onetier.retro_together.configuration;
+
+
+import com.onetier.retro_together.jwt.JwtFilter;
+import com.onetier.retro_together.jwt.TokenProvider;
+import com.onetier.retro_together.service.UserDetailsServiceImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.DefaultSecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+/**
+ * JwtSecurityConfig
+ */
+@RequiredArgsConstructor
+public class JwtSecurityConfiguration
+    extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
+
+  private final String SECRET_KEY;
+  private final TokenProvider tokenProvider;
+  private final UserDetailsServiceImpl userDetailsService;
+
+  @Override
+  public void configure(HttpSecurity httpSecurity) throws Exception {
+    JwtFilter customJwtFilter = new JwtFilter(SECRET_KEY, tokenProvider, userDetailsService);
+    httpSecurity.addFilterBefore(customJwtFilter, UsernamePasswordAuthenticationFilter.class);
+  }
+
+}
