+++ src/main/java/com/onetier/retro_together/configuration/SecurityConfiguration.java
@@ -1,0 +1,82 @@
+package com.onetier.retro_together.configuration;
+
+
+import com.onetier.retro_together.jwt.AccessDeniedHandlerException;
+import com.onetier.retro_together.jwt.AuthenticationEntryPointException;
+import com.onetier.retro_together.jwt.TokenProvider;
+import com.onetier.retro_together.service.UserDetailsServiceImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.autoconfigure.security.ConditionalOnDefaultWebSecurity;
+import org.springframework.boot.autoconfigure.security.SecurityProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+/**
+ * SecurityConfig
+ */
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+@ConditionalOnDefaultWebSecurity
+@ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
+public class SecurityConfiguration {
+
+  @Value("${jwt.secret}")
+  String SECRET_KEY;
+  private final TokenProvider tokenProvider;
+  private final UserDetailsServiceImpl userDetailsService;
+  private final AuthenticationEntryPointException authenticationEntryPointException;
+  private final AccessDeniedHandlerException accessDeniedHandlerException;
+
+  @Bean
+  public PasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
+
+  @Bean
+  @Order(SecurityProperties.BASIC_AUTH_ORDER)
+  public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    http.cors();
+
+    http.csrf().disable()
+
+        .exceptionHandling()
+        .authenticationEntryPoint(authenticationEntryPointException)
+        .accessDeniedHandler(accessDeniedHandlerException)
+
+        .and()
+        .sessionManagement()
+        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+
+        .and()
+        .authorizeRequests()
+        .antMatchers("/api/member/**").permitAll()
+        .antMatchers("/api/post/**").permitAll()
+        .antMatchers("/api/comment/**").permitAll()
+        .antMatchers( "/v2/api-docs",
+                      "/swagger-resources",
+                      "/swagger-resources/**",
+                      "/configuration/ui",
+                      "/configuration/security",
+                      "/swagger-ui.html",
+                      "/webjars/**",
+                      "/v3/api-docs/**",
+                      "/swagger-ui/**").permitAll()
+        .anyRequest().authenticated()
+
+        .and()
+        .apply(new JwtSecurityConfiguration(SECRET_KEY, tokenProvider, userDetailsService));
+
+    return http.build();
+  }
+
+}
