+++ src/main/java/com/onetier/retro_together/controller/request/LoginRequestDto.java
@@ -1,0 +1,23 @@
+package com.onetier.retro_together.controller.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+/**
+ * LoginRequestDto
+ */
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginRequestDto {
+
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String password;
+
+}
