+++ src/main/java/com/onetier/retro_together/controller/request/TokenDto.java
@@ -1,0 +1,22 @@
+package com.onetier.retro_together.controller.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+/**
+ * TokenDto
+ */
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TokenDto {
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+    private Long accessTokenExpiresIn;
+
+
+}
