+++ src/main/java/com/onetier/retro_together/controller/response/ResponseDto.java
@@ -1,0 +1,21 @@
+package com.onetier.retro_together.controller.response;
+import com.onetier.retro_together.domain.Error;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ResponseDto<T> {
+  private boolean success;
+  private T data;
+  private Error error;
+
+  public static <T> ResponseDto<T> success(T data) {
+    return new ResponseDto<>(true, data, null);
+  }
+
+  public static <T> ResponseDto<T> fail(String code, String message) {
+    return new ResponseDto<>(false, null, new Error(code, message));
+  }
+
+}
