+++ src/main/java/com/onetier/retro_together/domain/Error.java
@@ -1,0 +1,16 @@
+package com.onetier.retro_together.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Error {
+    private String field;
+    private String message;
+
+}
