+++ src/main/java/com/onetier/retro_together/repository/MemberRepository.java
@@ -1,0 +1,15 @@
+package com.onetier.retro_together.repository;
+
+import com.onetier.retro_together.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member,Long> {
+
+    Optional<Member> findById(Long id);
+    Optional<Member> findByEmail(String email);
+    Optional<Member> findByNickname(String nickname);
+
+
+}
