+++ src/main/java/com/onetier/retro_together/shared/Authority.java
@@ -1,0 +1,7 @@
+package com.onetier.retro_together.shared;
+
+public enum Authority {
+    ROLE_MEMBER,
+    ROLE_GUEST
+
+}
